# docs/openapi.yml
openapi: "3.0.3"
info:
  title: TopicPost API仕様書
  version: "1.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://www.topicpost.net
    description: 本番環境
  - url: https://stg.topicpost.net
    description: 検証環境

tags:
  - name: User
    description: ユーザー情報に関するAPI
  - name: Profile
    description: プロフィール情報に関するAPI

paths:
  /users:
    get:
      tags: [User]
      security:
        - bearerAuth: []
      operationId: users-get
      summary: ユーザー一覧を取得する
      description: |-
        ユーザー一覧を取得するAPIです
      responses:
        # 2xx系のレスポンス
        200:
          description: ユーザー一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserListResponse"

        # 4xx系のレスポンス
        400:
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        # 5xx系のレスポンス

    post:
      tags: [User]
      security:
        - bearerAuth: []
      operationId: users-post
      summary: ユーザーを作成する
      description: |-
        ユーザーを作成するAPIです
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        # 2xx系のレスポンス
        200:
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

        # 4xx系のレスポンス
        400:
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        # 5xx系のレスポンス

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: エラーメッセージ
          example: リクエストが不正です

    UserRequest:
      type: object
      required:
        - uid
      properties:
        uid:
          type: string
          format: uuid
          description: Supabaseが発行したUIDを入れる
          example: 123e4567-e89b-12d3-a456-426614174000

    User:
      type: object
      required:
        - id
        - uid
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーID
        uid:
          type: string
          format: uuid
          description: Supabaseが発行したUID
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時

    UserResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/User"

    UserListResponse:
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"